fix rating

+++ type/table/rating.go
@@ -9,7 +9,7 @@ type Rating struct {
 	Giver      *User      `gorm:"foreignKey:GiverId;references:UserId"`
 	ReceiverId *uint64    `gorm:"not null"` // FK ชี้ไปที่ User ที่ได้รับ rating
 	Receiver   *User      `gorm:"foreignKey:ReceiverId;references:UserId"`
-	Rating     *float64   `gorm:"not null"`
+	Rating     *uint64    `gorm:"not null"`
 	Score      *uint64    `gorm:"not null"`
 	NumRate    *uint64    `gorm:"not null"`
 	CreatedAt  *time.Time `gorm:"autoCreateTime"`
